feat: usePriceUtils, useTimeUtils 로직 분리 (추후 파일 이동)
