+++ src/pages/noticedetail/usePriceUtils.ts
@@ -1,0 +1,3 @@
+export const calculatePriceChange = (originalPay: number, currentPay: number): number => {
+  return ((currentPay - originalPay) / originalPay) * 100;
+};
