+++ src/shared/components/PostList/PostPrice/PostPrice.styled.ts
@@ -9,7 +9,7 @@ export const PostPriceContainer = styled.div`
   display: flex;
   justify-content: space-between;
   align-items: center;
-  /* gap: 8px; */
+  gap: 8px;
 
   @media (max-width: 744px) {
     flex-direction: column;
@@ -18,7 +18,7 @@ export const PostPriceContainer = styled.div`
 `;
 
 export const PostPrice = styled.h2`
-  /* width: 100%; */
+  width: 100%;
   font-size: 24px;
   font-weight: 700;
   letter-spacing: 0.48px;
@@ -30,7 +30,7 @@ export const PostPrice = styled.h2`
 
 export const PostPriceChange = styled.div<PostContainerProps>`
   display: flex;
-  /* width: 100%; */
+  width: 100%;
   height: 36px;
   padding: 12px 30px 12px 12px;
   align-items: center;
